requirements.txt
alembic==1.12.1
aniso8601==9.0.1
anyio==4.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
cffi==1.16.0
click==8.1.7
Django==3.2.*
fastapi==0.95.1
graphene==3.2.2
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
Mako==1.3.0
MarkupSafe==2.1.3
promise==2.3
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==1.10.13
PyJWT==2.7.0
python-dotenv==1.0.0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.11
sqlparse==0.4.4
starlette==0.26.1
starlette-graphene3==0.6.0
text-unidecode==1.3
typing_extensions==4.8.0
uvicorn==0.21.1
